Fix constant names with 'define'
The `define` function is interpreted by PHP-Parser as a regular
function and special transformations are needed to convert the
first argument into a namespaced constant name.

In https://github.com/phpDocumentor/phpDocumentor2/issues/729 it
was reported that this transformation resulted in erroneous 
names. With this commit we fix the way those define statements 
are interpreted and correctly extract the namespace and name.