Namespace aliases are not returned
When we try to retrieve the namespace aliases in the FileReflector we
have noticed that we did not receive any results despite that there are
namespace aliases present in a file.

During a previous refactoring the concept of a DocBlock Context was
introduced but the getters for namespace and namespace aliases were
not changed to retrieve their information from that context class.

In this commit I have altered the FileReflector to use the context and
I have cleaned up a few parts.

This change fixes https://github.com/phpDocumentor/phpDocumentor2/issues/1313.