Merged Descriptors into Reflection library
In an attempt to improve the performance of phpDocumentor we have
merged the Descriptor DTO layer of phpDocumentor into the Reflection
library. The goal is to write a series of Assemblers that will directly
convert the output of PHP-Parser into Descriptors and skipping the
Reflection objects.

In addition to the above I have updated PHP-Parser to version 1.0 and
replaced all non-namespaced code to use their namespaced variants so
that we are ready once the non-namespaced aliases are removed.