Upgrades phpParser to version ^2.0
This will allow us to use the new php7 features of phpParser.
I added a optional parameter to the NodesFactory so we can even specify the
prefered parser type. This could be added to the configuration of the project
so phpdocumentor knows which version of php it is dealing with.

This will allow us to add optional validation based on the php version?