A failure in parsing an FQSEN causes the whole project to crash
In this change I catch any exception occuring during the interpretation of a
file so that phpDocumentor won't stop generating documentation when a single
file is corrupt or contains a critical error.

The expected behaviour is that this issue is ignored, or logged, and that
parsing should continue for other files.