Support for Typed Properties
So far this library did not support Typed Properties (a PHP 7.4 feature)
and to stay relevant we wanted to introduce this. Nikita Popov's library
made this rather simple but we did find an issues with Union types from
that library.

As such, in this change I included a new factory to convert the types of
php parser into our own Type system.