Remove custom PrettyPrinter
In a previous version of phpDocumentor we had introduced a 'hack' to
more easily cope with escaping in XML by stripping quotes and other
formatting from the pretty printing of strings.

Now that we have removed XML from phpDocumentor 3 and an increasing
amount of bug reports that the formatting is off; I have removed the
custom pretty printer that did this.