Improve factory strategies
The factory strategies are now executed by priority this allows
users to inject their own strategies with high priority to overrule
the default behavior of the library.

The stategies are updated to set their own result to the context of
creation. This allows strategies to create a property and a method
for example which is required to support property promotion in php8.