Fix #233 inline defined const
Defines in a method or function are allowed in php. However they are added
to the global scope. As this library now processes the full method body to be
able to add more information into the reflected code, this caused issues.

As the defines are added to the global scope of the execution they are basically
defined at a file level when a define was executed. This is the best we can get right now to
reflect what the application would do at runtime.