Started reworking expression strings into Expression objects
Values of constants, named arguments, etc, should be parsed to include
types and FQSENs that can be used in phpDocumentor to display links
with. Since these are now simple strings, I have refactored this into a
template string system with placeholder values that a consumer could
replace with its own values.