Add integration tests for Expressions, and fail
As part of checking whether everything works as expected I am adding
integration tests, and find out that the pretty printer can be a bit
greedy. This is exemplified in the ConstructorPromotionTest where I want
to pass a reference to a local class constant, but the ExpressionPrinter
picks up on the class reference and does not resolve it.

This is a major issue, and more work is needed to

a) parse the whole FQSEN and use that
b) find out how to resolve references, such as self

Another test is needed to see if it resolves relative references as
well, as the Standard PrettyPrinter code gives me the idea it does not.